Moved paging into container
